Adicionado a funcao de validar data da inscricao

+++ app/Http/Controllers/ServidorController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Carbon\Carbon;
 use App\Modalidades;
 use App\Inscricao;
 
@@ -46,6 +47,18 @@ class ServidorController extends Controller
         return view('inscricao1', compact('modalidades','inscricoes','rmr'));
     }
     public function inscricao1_store(Request $request){
+        /* Valida se está no periodo de inscrições */
+        $data_inicial_insc = Carbon::createFromFormat('Y-m-d H:i:s', env('DT_INICIO_INSC', '2018-08-20 00:00:00'));
+        $data_final_insc = Carbon::createFromFormat('Y-m-d H:i:s', env('DT_TERMINO_INSC', '2018-08-24 23:59:59'));
+        $agora = Carbon::now();
+        if ($agora->lt($data_inicial_insc)){
+            $msg_erro = 'As inscrições ainda não estão abertas.';
+            return redirect()->route('inscricao1')->withErrors([$msg_erro]);
+        }
+        if ($agora->gt($data_final_insc)){
+            $msg_erro = 'As inscrições estão encerradas.';
+            return redirect()->route('inscricao1')->withErrors([$msg_erro]);
+        }
         DB::beginTransaction();
         try{
             Inscricao::where('user_id', \Auth::user()->id)->delete();
